refactor(frontend): decompose QuizEditor into editor-scoped modules
QuizEditor.tsx shrank from 625 lines to 151 lines. The component now
acts as a thin orchestrator around a draft-state hook and prop-driven
presentational components under `components/quiz/editor/`:

- types.ts            — DraftOption/DraftQuestion + factories (uid,
                        makeDefaultQuestion, makeDefaultOption,
                        makeTrueFalseOptions)
- useQuizDraft.ts     — hook encapsulating load + all question/option
                        CRUD reducers (add/remove/move/update)
- QuizHeaderFields.tsx — title/description/passing score/max attempts
- QuestionCard.tsx    — single question with MCQ/T-F/SA/essay variants
- QuizEditView.tsx    — header + question list + save/delete actions
- ModeToggle.tsx      — Edit ↔ Submissions toggle
- index.ts            — barrel exports

No functional/behavioral changes. tsc, eslint, vitest (85/85) and
`npm run build` all pass.

Made-with: Cursor
